src/requirements.txt
streamlit
pandas
requests
googletrans==4.0.0-rc1
python-dotenv
